projects/mini-hacker-pi/src/shell/requirements.txt
pillow
RPi.GPIO
spidev
st7789